Refactor a few things
